pyproject.toml
[tool.poetry]
name = "agent_bus"
version = "0.1.0"
description = "Multi-agent SWE engineering system with distributed GPU compute"
authors = ["agent_bus team"]
readme = "README.md"
python = "^3.10"

[tool.poetry.scripts]
agent-bus-skills = "src.cli:main"
agent-bus-memory = "src.cli_memory:main"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
redis = "^5.0.0"
psycopg2-binary = "^2.9.9"
anthropic = "^0.18.0"
httpx = "<0.28"
pydantic = "^2.5.0"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.0"
chromadb = "^0.4.22"
sentence-transformers = "^2.3.1"
kubernetes = "^29.0.0"
prometheus-client = "^0.19.0"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
packaging = "^23.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.1.0"
ruff = "^0.2.0"
mypy = "^1.8.0"

[tool.poetry.group.ml]
optional = true

[tool.poetry.group.ml.dependencies]
torch = "^2.2.0"
torchvision = "^0.17.0"
numpy = "^1.26.0"
scikit-learn = "^1.4.0"
opencv-python = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.per-file-ignores]
"tests/**/*.py" = ["E402"]  # Allow imports after path manipulation in tests

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
